Continued Improvements to MCP
Steal money is now being controlled by MCP, and will continue to launch at new targets when memory is available.

Nuke servers is now being controller by MCP, with each run command in a separate file, so it takes less resources to run

Migrated the rest of the useful scripts to the MCP pattern, still work to be done to parameterize things better.